Change version
Revised README.md for readability

The following is a sample of my work for your consideration.
I don't know anything about ML, but I find this product fascinating.
I'm speaking with Unusual tomorrow about a Technical writing position at
your company, and thought I'd put this in.

129-132: The note instructs the user on how to use the --force flag.
However the example that follows also includes the --template option.
I'm not sure how the template option plays into it here. I could run the
rest of these commands, but the template part confuses me. What is this
template I'm pointing to and it doesn't seem like it belongs in this
note section.

160-168:

Do I need an existing docker.yml for this to work?
See output

   ➜  ~ INIT_COMMAND='--init' docker-compose up -d
   ERROR:
           Can't find a suitable configuration file in this directory or any
           parent. Are you in the right directory?

           Supported filenames: docker-compose.yml, docker-compose.yaml

175: This may be me, but I'm not sure what this file is or where to find it.
Is it the docker config, I'm new to docker, but I've got the commands to run fine up to this point.

203: Unsure of full Functionality, so review the Collaborative section
here and add some more detail. Are we talking just comments, suggesting
changes, workflow approvals?

204: listing the specific attributes "_images _audios" etc didn't belong
in a feature list. So they ahve been removed

206: This sentence feels like it needs more. Does the interfvace work on mobile? what degree of control do I have through the mobile interface.
I don't write marketing copy, so I'm inclined to remove this whole line, because it seems to me that just saying "mobile support" should belong in a feature list.

229:
This section taught me that I don't know certain things about ML, so take note of that.
For Diarization, I think Diarisation is right, I'm new to the term, but I'm willing to be good money that spelling with "s" is the UK spelling. So I suggest using Diarization with a z instead. I could be wrong.

230: I added *audio-based* emotion recognition despite it being redundant.
Emotion recognition was the only one in tat series that didn't imply audio in the title .
I image there is video based emtion reocognition,
which sound very different to me that it is worth calling out again by switcing to *Audio-based* emotion recognition.