fix: LEAP-361: Use named exports for SVG
This will allow to use new version of svg imports used in monorepo.
So transition to monorepo will be smoother and this will unblock LS develop right now.
The reason is that monorepo webpack uses another settings of SVGR and uses named exports by default,
they can be enabled with option `exportType: named` in current version.