Added StrokeTool to be a stroke tool base class that handles resizing of cursor correctly.
This also rechecks if the cursor is set to default after hovering over region, and will reset the cursor accordingly.

Now correctly shows the cursor size to the tools strokeWidth.
Since the cursor icon can only be a max of 128x128 pixels, any larger is denoted by a dashed circle instead of a full circle.
Could fix this issue by adding an element that follows the mouse but for now this is good enough.