Added a visibility option.
When selecting layers all other layers will be hidden.
hidden implementation can be implemented under regions of the respective object.

brush region has setVisiblity to set the opacity and any other region by default sets itself to hidden or not hidden.

Added a toggle to turn this on and off.