Added StrokeTool under mixins to parent class the eraser and brush for cursor functionality and extra annotation functionality.
StrokeTool has a fine stroke palette, a list of numbers can be given in order of big to small and when the stroke is smaller than the biggest in the list it will be set to any of the predefined brush strokes. This allows for more finer control of the brush when zooming in and using the increase/decrease brush size command.
Before the brush would increment in 5's but now by default it will increment/decrement from 10 to 9, 8, 7, 5, 4.5, 4, 3.5, 3, 2.5, 2, 1.5, 1. Instead of 10, 5, 1 with default increment of 5 and clamp.

The cursor icon for the stroke tools is also updated to be as big as 128 pixels, and after that the cursor turns into a dotted circle to show that the brush is too big for the cursor as the cursor is capped at 128x128 pixels. This can be fixed by using a div that follows the mouse.

The cursor also takes in the zoom scale to correctly denote the brush size at different zooms.

The addPoint has the Math.round removed when adding brush stroke points as they did not correctly line upo with the cursor position.

Added an active brush opacity for the stroke tool that will update the tools opacity when drawing or erasing. This is set as an extra slider next to the size slider and is stored in local storage. The feature allows the user to be able to set the opacity to near 0 when erasing to correctly line up the are to erase without overshooting as it will be easier to see under the existing paint brush. Same with the brush, the opacity for each is seperate and stored locally. This feature can be turned on in settings.

Added a toggle in settings that removes the auto annotation tooltip at the top of the images. Toggleable in settings.

Updated region store to toggle visiblity of non selected regions. Just like the brush opacity, but this will turn off all areas that are not currently selected when selecting areas. This can be toggled on or off at the settings tab.

Modifed the canvas create circle to allow for dashed circles.
Removed toggles for features at the bottom of the application and moved them to the settings tab.

Added corresponding attributes to settings store which are saved in local storage.