fix: RND-97: Support for multiple ML backend predictions per task (#5982)
- Support for multiple predictions response from ML backend. Current
version only supports ML backend /predict response with the following
payload: `"results": [{}, {}, ...]` where it expects the list of the
predictions of the same length as the number of submitted tasks. Now it
can optionally handle the case when ML backend responds with `"results":
[ [{}, {}], ...]` with each item represents the list of Predictions per
task (e.g. with different model versions)
- refactor multiple calls to ML backend to simplify the code (most of
the changes)
- added tests for single and multiple predictions, on task opening as
using "actions" with sdk

---------

Co-authored-by: nik <nik@heartex.net>
Co-authored-by: makseq-ubnt <makseq@gmail.com>