fix: TRIAG-600: Add missing storage title field to Redis config (#5985)
[Re-created using `gh` command line tool for CI (mirror #5552]

This commit adds a required missing field to the Redis storage form --
the storage title. Without this field, the UI will throw an exception
when trying to display the saved configuration's card. See
[Slack](https://label-studio.slack.com/archives/CQ8LYQ0J2/p1709796798556099)
for details.

Also fix the port showing up as `undefined` on the frontend's storage
card. We'll need to make a similar change in LSE.

---------

Co-authored-by: Justin Ziniel <ziniel.1@osu.edu>
Co-authored-by: jombooth <jombooth@users.noreply.github.com>
Co-authored-by: robot-ci-heartex <robot-ci-heartex@users.noreply.github.com>